Adding lastUpdate and character_last_update to optimise the Vector Store

+++ app.py
@@ -12,6 +12,7 @@ from langchain.memory import ConversationBufferWindowMemory
 from langchain.vectorstores.base import VectorStoreRetriever
 from langchain_chroma import Chroma
 from langchain_community.document_loaders import DataFrameLoader
+from langchain_community.vectorstores.utils import filter_complex_metadata
 from langchain_core.output_parsers import StrOutputParser
 from langchain_core.prompts import ChatPromptTemplate
 from langchain_core.runnables import RunnableConfig
@@ -33,6 +34,9 @@ def load_documents(df, page_content_column: str):
         if not doc.metadata.get(page_content_column):
             # Add or update the page_content_column in metadata
             doc.metadata[page_content_column] = doc.page_content
+            # Handle timestamp metadata
+            time_stamp = doc.metadata.get("character_last_update")
+            doc.metadata["character_last_update"] = str(time_stamp) if time_stamp else ""
     return documents
 
 
@@ -47,34 +51,44 @@ def load_vectordb(init: bool = False):
     global vectordb
     VECTORDB_FOLDER = ".vectordb"
     df_character = user_session.get("current_character")
-    if not init and vectordb is None:
+    character_id = int(df_character.id.iloc[0])
+    last_updated_db = df_character["lastUpdated"].iloc[0]
+    if os.path.exists(VECTORDB_FOLDER):
         vectordb = Chroma(
             embedding_function=init_embedding_function(),
             persist_directory=VECTORDB_FOLDER,
             client_settings=Settings(anonymized_telemetry=False, is_persistent=True),
         )
-        if not vectordb.get()["ids"]:
+
+    if not init:
+        # Retrieve the lastUpdated value from the vector store
+        last_updated_vector_store = None
+        try:
+            last_updated_docs = vectordb.get(where={"character_id": int(character_id)})
+            if last_updated_docs and last_updated_docs["metadatas"][0]["character_last_update"]:
+                last_updated_vector_store = last_updated_docs["metadatas"][0]["character_last_update"]
+        except Exception as e:
+            logger.error(f"Error retrieving lastUpdated from vector store: {e}")
+
+        if last_updated_vector_store != str(last_updated_db):
+            logger.info("lastUpdated value has changed, re-initializing vector store")
             init = True
         else:
-            logger.info(f"Vector DB loaded")
+            logger.info(f"Vector DB loaded, lastUpdated matches")
+
     if init:
-        if os.path.exists(VECTORDB_FOLDER):
-            vectordb = Chroma(
-                embedding_function=init_embedding_function(),
-                persist_directory=VECTORDB_FOLDER,
-                client_settings=Settings(anonymized_telemetry=False, is_persistent=True),
-            )
-            character_docs = vectordb.get(where={"character_id": int(df_character.id.iloc[0])})
+        if os.path.exists(VECTORDB_FOLDER) and vectordb:
+            character_docs = vectordb.get(where={"character_id": int(character_id)})
             logger.info(f"Character docs: {character_docs}")
             if character_docs.get("ids"):
                 vectordb.delete(character_docs.get("ids"))
-                logger.info(f"Deleted existing Vector DB for character {df_character.id.iloc[0]}")
+                logger.info(f"Deleted existing Vector DB for character {character_id}")
             else:
-                logger.info(f"No existing Vector DB found for character {df_character.id.iloc[0]}")
+                logger.info(f"No existing Vector DB found for character {character_id}")
         # Load latest data from Postgres DB for the current character
-        docs = load_documents(db.load_dialogues(int(df_character.id.iloc[0])), page_content_column="utterance")
+        docs = load_documents(db.load_dialogues(int(character_id)), page_content_column="utterance")
         vectordb = Chroma.from_documents(  # Create a new Vector DB from the loaded documents
-            documents=docs,  # Load dialogue utterances
+            documents=filter_complex_metadata(docs),  # Load dialogue utterances
             embedding=init_embedding_function(),  # Initialize embedding function
             persist_directory=VECTORDB_FOLDER,
             client_settings=Settings(anonymized_telemetry=False, is_persistent=True),
@@ -166,7 +180,7 @@ async def start():
         logger.error(
             "No available characters found in df_character. Please check the Postgres DB for the 'Character' table.")
         return
-    load_vectordb(True)
+    load_vectordb()
 
     await set_character()
